==> analysis/04_check_subject_exclusions.Rmd <==
---
title: "Check Subject Exclusions"
date: "`r Sys.Date()`"
author: "Kyle Kurkela"
output: 
  html_document: 
    toc: yes
    toc_float: yes
    code_folding: hide
    df_print: paged
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)

library(tidyverse)
library(rmarkdown)
library(ggridges)

tidy.enc   <- read_rds('tidy_data/tidy_enc.rds')
tidy.ret   <- read_rds('tidy_data/tidy_ret.rds')
tidy.bds   <- read_rds('tidy_data/tidy_bds.rds')
tidy.catch <- read_rds('tidy_data/tidy_catch.rds')
```

## Exclusion Criterion 1

Participants who are statistical outliers (3 standard deviations below the mean)
for reaction time for any 1 of our tasks.

### Encoding RTs

Table is sorted by session (`session`) and the absolute value of the z score of their average reaction time for encoding trials (specifically, how long it took them to submit their imagination success rating; `mean_rt_slider`). Does NOT include catch trials) calculated within session (`zscore`). See code.  

```{r}
# encoding
tidy.enc %>%
  # calculate the mean RT for the imagination success slider for each subject/session
  group_by(subject_id, session) %>%
  summarise(mean_rt_slider = mean(rt_slider), .groups = 'drop') %>%
  # calculate a zscore within session
  group_by(session) %>%
  mutate(zscore = as.double(scale(mean_rt_slider))) %>%
  # exclude subjects faster than 3 SDs
  ungroup() %>%
  mutate(exclude_enc = zscore < -3) %>%
  # rts are in milliseconds. Convert to seconds
  mutate(mean_rt_slider = mean_rt_slider / 1000) %>%
  # arrange by session and descending absolute value zscore
  arrange(session, desc(abs(zscore))) %>%
  paged_table()
```

```{r, warning=FALSE, message=FALSE}
tidy.enc %>%
  # rt milliseconds --> seconds
  mutate(rt_slider = rt_slider / 1000) %>%
  # a density ridges plot. See ggridges package.
  ggplot(aes(x = rt_slider, y = subject_id)) +
  geom_point(shape = '|', size = 1.5) +
  scale_x_continuous(breaks = seq(0, 300, 50)) +
  labs(x = 'slider_rt (s)', title = 'Encoding Slider Response', subtitle = 'RT Distributions', y = 'subjects') +
  facet_grid(~session) +
  theme(axis.text.y = element_blank(), axis.ticks.y = element_blank())
```

### Cued Recall RTs

Table is sorted by session (`session`) and the absolute value of the z score of their average reaction time for cued recall trials calculated within session (`zscore`). See code.  

```{r}
# cued recall
tidy.ret %>%
  # calculate the mean RT for the imagination success slider for each subject/session
  group_by(subject_id, session) %>%
  summarise(mean_rt_cued_recall = mean(rt), .groups = 'drop') %>%
  # calculate a zscore within session
  group_by(session) %>%
  mutate(zscore = as.double(scale(mean_rt_cued_recall))) %>%
  # exclude subjects faster than 3 SDs
  ungroup() %>%
  mutate(exclude_ret = zscore < -3) %>%
  # arrange by session and descending absolute value zscore
  arrange(session, desc(abs(zscore))) %>%
  paged_table()
```

```{r, warning=FALSE,message=FALSE}
tidy.ret %>%
  # rt milliseconds --> seconds
  mutate(rt = rt / 1000) %>%
  # a density ridges plot. See ggridges package.
  ggplot(aes(x = rt, y = subject_id)) +
  geom_vline(color = 'red', xintercept = 600, linetype = 'dotted') +
  geom_point(shape = '|', size = 1.5) +
  scale_x_continuous(breaks = seq(0,700, 100)) +
  labs(x = 'slider_rt (s)', title = 'Cued Recall', 
       subtitle = 'Submit RT Distributions', y = 'subjects', 
       caption = 'dashes = trials; red dashed = 10 minutes.') +
  facet_grid(~session) +
  theme(axis.text.y = element_blank(), axis.ticks.y = element_blank())
```

### BDS RTs

Table is sorted by session (`session`) and the absolute value of the z score of their average reaction time for BDS trials calculated within session (`zscore`). See code.  

```{r}
# bds
tidy.bds %>%
  # calculate the mean RT for the imagination success slider for each subject/session
  group_by(subject_id, session) %>%
  summarise(mean_rt_bds = mean(rt), .groups = 'drop') %>%
  # calculate a zscore within session
  group_by(session) %>%
  mutate(zscore = as.double(scale(mean_rt_bds))) %>%
  # exclude subjects faster than 3 SDs
  ungroup() %>%
  mutate(exclude_bds = zscore < -3) %>%
  # arrange by session and descending absolute value zscore
  arrange(session, desc(abs(zscore))) %>%
  paged_table()
```

```{r}
tidy.bds %>%
  # rt milliseconds --> seconds
  mutate(rt = rt / 1000) %>%
  # a density ridges plot. See ggridges package.
  ggplot(aes(x = rt, y = subject_id)) +
  geom_point(shape = '|', size = 1.5) +
  scale_x_continuous(breaks = seq(0,300,50)) +
  labs(x = 'reaction time (s)', 
       title = 'Backwards Digit Span', 
       subtitle = 'RT Distributions', 
       y = 'subjects', caption = 'dashes = trials') +
  facet_grid(~session) +
  theme(axis.text.y = element_blank(), axis.ticks.y = element_blank())
```

## Exclusion Criterion 2

Participants who objectively demonstrated clear low-effort throughout the experiment. Specifically participants are missing responses from more than 90% of cued recall trials. Table is sorted by session (`session`) and proportion of cued recall response prompts left blank (`prop_blank`).    

```{r}
tidy.ret %>%
  # data from wide --> long
  pivot_longer(cols = starts_with('ret_resp'), names_to = 'resp_num', values_to = 'text_resp') %>%
  # calculate proportion left blank for each subject/session
  group_by(subject_id, session) %>%
  summarise(nblank = sum(text_resp == ""), total = n(), .groups = 'drop') %>%
  mutate(prop_blank = nblank / total) %>%
  # exclude IFF proportion blank > 90%
  mutate(exclude = prop_blank > .9) %>%
  # sort table by session, proportion blank
  arrange(session, desc(prop_blank)) -> propBlank.df

paged_table(propBlank.df)
```

## Exclusion Criterion 3

Participants who objectively demonstrated clear low-effort throughout the experiment. Specifically participants used the same response on the slider scale for every single encoding trial. Table below is sorted by session (`session`) and the standard deviation of their slider responses.  

```{r}
tidy.enc %>%
  # standard deviation of slider response for each subject/session
  group_by(subject_id, session) %>%
  summarise(sd_slider = sd(response_slider), .groups = 'drop') %>%
  # exclude subjects who have a standard deviation == 0 -- they had the same response every time
  mutate(exclude = sd_slider == 0) %>%
  # sort by session, sd of the slider
  arrange(session, sd_slider) %>%
  paged_table()
```

## Exclusion Criterion 4

Participants who objectively demonstrated clear low-effort throughout the experiment. Specifically, they responded with nonsense on the catch trials. The table below is sorted by session (confusingly labeled `session`) and average length, in characters, of their responses (`aveRespLength`). See code.

Session 1 only:

```{r}
# catch trials
tidy.catch %>%
  filter(session == 'session1') %>%
  # calculate number of characters for each response
  mutate(respLength = str_length(response)) %>%
  # calculate average response length for each subject/session
  group_by(subject_id) %>%
  mutate(avgRespLength = mean(respLength)) %>% 
  arrange(avgRespLength) %>%
  dplyr::select(subject_id, session, response) %>%
  paged_table()
```

Session 2 only:

```{r}
# catch trials
tidy.catch %>%
  filter(session == 'session2') %>%
  # calculate number of characters for each response
  mutate(respLength = str_length(response)) %>%
  # calculate average response length for each subject/session
  group_by(subject_id) %>%
  mutate(avgRespLength = mean(respLength)) %>% 
  arrange(avgRespLength) %>%
  dplyr::select(subject_id, session, response) %>%
  paged_table()
```

## Suspect Cued Recall Response Lengths

Identify individuals with unusually long response lengths.

```{r, warning=FALSE, message=FALSE}
tidy.ret %>% 
  mutate(respLength = str_length(ret_resp_1) + str_length(ret_resp_2)) %>% 
  group_by(subject_id, session) %>% 
  summarise(across(respLength, max), .groups = 'drop') %>% 
  arrange(desc(respLength))
```
