==> analysis/04_time_spent_analysis.Rmd <==
---
title: "Timespent Analysis"
date: "`r Sys.Date()`"
author: "Kyle Kurkela"
output: 
  html_document: 
    toc: yes
    toc_float: yes
    code_folding: hide
    df_print: paged
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)

library(tidyverse)

df <- read_rds('tidy_data/compiled_experiment.rds')
```

See code chunk below for Kyle's tidying process

```{r tidy}
df %>%
  filter(phase %in% c('enc', 'bds', 'ret')) %>%
  arrange(subject_id, session, trial_index) %>%
  group_by(subject_id, session, phase) %>%
  slice(c(1,n())) %>%
  group_by(subject_id, session, phase) %>%
  summarise(time_spend = diff(time_elapsed), trial_index = max(trial_index), .groups = 'drop') -> timespent.task

df %>%
  filter(!phase %in% c('bds', 'enc', 'ret')) %>%
  arrange(subject_id, session, trial_index) %>%
  dplyr::select(subject_id, session, trial_index, phase, rt) %>%
  filter(!is.na(rt)) %>%
  mutate(rt = as.double(rt)) %>%
  rename(time_spend = rt) -> timespent.all.else

timespent.all.else %>%
  bind_rows(., timespent.task) %>%
  arrange(subject_id, trial_index) %>%
  mutate(time_spend = time_spend / 1000 / 60) -> tidy.timespend

set.seed(321)
SelectColors <- sample(grDevices::rainbow(n = 14), 14)
```

## How much time did participants send in each phase of the experiment?

```{r}
tidy.timespend %>%
  arrange(session, subject_id, trial_index) %>%
  rmarkdown::paged_table()
```

A Pageable Table to explore. Where:

- `subject_id` = arbitrary label given by Prolific.  
- `trial_index` = arbitrary numeric label given to each phase, designed to sort them in chronological order.  
- `phase` = arbitrary labels given to the different parts of the experiment. Should be easy enough to figure out.  
- `time_spend` = estimated amount of time (in minutes) spend on each phase.  
- `session` = session 1 or session 2.  

## Which phases were most variable?

Session 1:  

```{r}
tidy.timespend %>%
  filter(session == 'session1') %>%
  group_by(phase) %>% 
  summarize(across(time_spend, sd), .groups = 'drop') %>%
  rename(time_spend_sd = time_spend) %>%
  arrange(desc(time_spend_sd)) %>%
  rmarkdown::paged_table()
```

Session 2:  

```{r}
tidy.timespend %>%
  filter(session == 'session2') %>%
  group_by(phase) %>% 
  summarize(across(time_spend, sd), .groups = 'drop') %>%
  rename(time_spend_sd = time_spend) %>%
  arrange(desc(time_spend_sd)) %>%
  rmarkdown::paged_table()
```

## Can we make a pretty visual?

```{r}
tidy.timespend %>%
  filter(session == 'session1') %>%
  mutate(subject_id = fct_reorder(.f = subject_id, .x = time_spend, .fun = sum)) %>%
  ggplot(aes(x = subject_id, y = time_spend)) +
  geom_col(aes(fill = phase), color = 'black') +
  coord_flip() +
  scale_y_continuous(breaks = seq(0,90,30)) +
  expand_limits(y = c(NA,120)) +
  scale_fill_manual(values = SelectColors) +
  labs(y = 'time spent (min)', 
       x = 'subject', 
       title = 'Where did subject spend their time?',
       subtitle = 'Session 1',
       caption = 'Colors sorted in chronological order from bottom to top.') +
  theme(axis.text.y = element_blank(), axis.ticks.y = element_blank())
```

```{r}
tidy.timespend %>%
  filter(session == 'session2') %>%
  mutate(subject_id = fct_reorder(.f = subject_id, .x = time_spend, .fun = sum)) %>%
  ggplot(aes(x = subject_id, y = time_spend)) +
  geom_col(aes(fill = phase), color = 'black') +
  coord_flip() +
  scale_y_continuous(breaks = seq(0,120,30)) +
  expand_limits(y = c(NA,120)) +
  scale_fill_manual(values = SelectColors) +
  labs(y = 'time spent (min)', 
       x = 'subject', 
       title = 'Where did subject spend their time?',
       subtitle = 'Session 2',
       caption = 'Colors sorted in chronological order from bottom to top.') +
  theme(axis.text.y = element_blank(), axis.ticks.y = element_blank())
```

Alternate Visual

```{r, warning=FALSE,message=FALSE}
library(ggbeeswarm)

tidy.timespend %>%
  group_by(subject_id, session) %>%
  summarise(across(time_spend, sum), .groups = 'drop') %>%
  group_by(session) %>%
  summarise(across(time_spend, median)) %>%
  mutate(time_spend = round(time_spend, digits = 2)) -> out.df

tidy.timespend %>%
  ggplot(aes(x = '', y = time_spend, group = subject_id)) +
  geom_quasirandom(stat = 'summary', fun = sum) +
  geom_hline(data = out.df, aes(yintercept = time_spend), color = 'red') +
  geom_label(data = out.df, aes(label = time_spend, y = time_spend, group = NULL)) +
  expand_limits(y = c(NA, 120)) +
  scale_y_continuous(breaks = seq(20,120,20)) +
  facet_grid(~session) +
  theme(axis.title.x = element_blank(), axis.ticks.x = element_blank()) +
  labs(caption  = 'dots = subjects, red line = median time (min)', 
       y = 'time spent (min)',
       title    = 'Total Time Spent on Experiment')
```

### Zoom in on retieval

Retrieval was the most variable.

```{r, warning=FALSE, message=FALSE}
tidy.timespend %>%
  filter(phase == 'ret') %>%
  ggplot(aes(x = '', y = time_spend, group = subject_id)) +  
  geom_quasirandom(stat = 'summary', fun = sum) +
  scale_y_continuous(breaks = seq(0,60,10)) +
  expand_limits(y = c(0,NA)) +
  facet_grid(~session) +
  theme(axis.title.x = element_blank(), axis.ticks.x = element_blank()) +
  labs(caption  = 'dots = subjects',
       y = 'time spent (min)',
       title    = 'How much time did participants spent on retrieval?')
```

### Zoom in on the Welcome Screen

```{r, warning=FALSE, message=FALSE}
tidy.timespend %>%
  filter(phase == 'welcome_screen') %>%
  ggplot(aes(x = '', y = time_spend, group = subject_id)) +  
  geom_quasirandom(stat = 'summary', fun = sum) +
  scale_y_continuous(breaks = seq(0,30,5)) +
  expand_limits(y = c(0,30)) +
  facet_grid(~session) +
  theme(axis.title.x = element_blank(), axis.ticks.x = element_blank()) +
  labs(caption  = 'dots = subjects', 
       y = 'time spent (min)',
       title    = 'How much time did participants spent on the welcome screen?')
```

### Zoom in on instr_pre_bds

```{r, warning=FALSE, message=FALSE}
tidy.timespend %>%
  filter(phase == 'instr_pre_bds') %>%
  ggplot(aes(x = '', y = time_spend, group = subject_id)) +  
  geom_quasirandom(stat = 'summary', fun = sum) +
  scale_y_continuous(breaks = seq(0,20,5)) +
  expand_limits(y = c(0,20)) +
  facet_grid(~session) +
  theme(axis.title.x = element_blank(), axis.ticks.x = element_blank()) +
  labs(caption  = 'dots = subjects',
       y = 'time spent (min)',
       title    = 'How much time did participants spent on the instructions screen before BDS?')
```

### Zoom in on encoding

```{r, warning=FALSE, message=FALSE}
tidy.timespend %>%
  filter(phase == 'enc') %>%
  ggplot(aes(x = '', y = time_spend, group = subject_id)) +  
  geom_quasirandom(stat = 'summary', fun = sum) +
  scale_y_continuous(breaks = seq(5,40,5)) +
  expand_limits(y = c(5,30)) +
  facet_grid(~session) +
  theme(axis.title.x = element_blank(), axis.ticks.x = element_blank()) +
  labs(caption  = 'dots = subjects',
       y = 'time spent (min)',
       title    = 'How much time did participants spent on encoding?')
```

## How long do participants spend completing the surveys?

```{r}
tidy.timespend %>%
  filter(phase %in% c('sam', 'iri', 'vviq')) %>%
  mutate(subject_id = fct_reorder(.f = subject_id, .x = time_spend, .fun = sum)) %>%
  ggplot(aes(x = subject_id, y = time_spend, fill = phase)) +
  geom_col(color = 'black') +
  scale_y_continuous(breaks = seq(0,30,3)) +
  scale_fill_manual(values = SelectColors[1:3]) +
  coord_flip() +
  #expand_limits(y = c(0,12)) +
  labs(title = 'How long did participants spend on the surveys?',
       y = 'time spent (min)') +
  theme(axis.text.y = element_blank(), axis.ticks.y = element_blank())
```
