==> analysis/06_secondary_exclusions.Rmd <==
---
title: "Secondary Exclusion Criteria"
date: "`r Sys.Date()`"
author: "Kyle Kurkela"
output:
  html_document: 
    toc: yes
    toc_float: yes
    code_folding: hide
    df_print: paged
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)

library(tidyverse)

df.interaction <- read_rds('tidy_data/compiled_interaction.rds')
df.experiment <- read_rds('tidy_data/compiled_experiment.rds')
df.graded <- read_rds('tidy_data/graded_df.rds')

df.sam <- read_rds('tidy_data/tidy_sam.rds')
df.iri <- read_rds('tidy_data/tidy_iri.rds')
df.vviq <- read_rds('tidy_data/tidy_vviq.rds')
```

## Criterion 1: RT Slower than 10 minutes

Reaction time on any 1 trial during encoding/retrieval is slower than 5 minutes.

```{r, warning=FALSE}
# coerce rt to double
# remove events without a rt
df.experiment %>%
  filter(!is.na(rt)) %>%
  group_by(subject_id, session, phase) %>%
  summarise(exclude = any(rt > 10*60*1000), .groups = 'drop') %>%
  filter(phase %in% c('enc', 'ret')) %>%
  pivot_wider(names_from = session, values_from = exclude) %>%
  rmarkdown::paged_table()
```

```{r}
df.experiment %>%
  filter(!is.na(rt)) %>%
  filter(phase == 'enc') %>%
  mutate(highlight = rt > 10*60*1000) %>%
  mutate(rt = rt/1000/60) %>%
  ggplot(aes(x = rt, y = subject_id)) +
  geom_vline(xintercept = 10, color = 'red', linetype = 'dotted') +
  geom_point(shape = "|", size = 1.5) +
  facet_grid(~session) +
  scale_color_manual(values = c('black', 'red')) +
  theme_light() +
  guides(color = "none") +
  theme(panel.grid.minor.x = element_blank(), 
        axis.text.y = element_blank(), 
        axis.ticks.y = element_blank()) +
  scale_x_continuous(breaks = c(0,1,5,10,15)) +
  labs(x = 'rt (min)', caption = 'dashes = trials. red line is 10 minutes', title = 'Encoding')
```

```{r}
df.experiment %>%
  filter(!is.na(rt)) %>%
  filter(phase == 'ret') %>%
  mutate(highlight = rt > 10*60*1000) %>%
  mutate(rt = rt/1000/60) %>%
  ggplot(aes(x = rt, y = subject_id)) +
  geom_point(shape = "|", size = 1.5) +
  facet_grid(~session) +
  geom_vline(xintercept = 10, color = 'red', linetype = 'dotted') +
  scale_color_manual(values = c('black', 'red')) +
  theme_light() +
  guides(color = "none") +
  theme(panel.grid.minor.x = element_blank(), 
        axis.text.y = element_blank(), 
        axis.ticks.y = element_blank()) +
  scale_x_continuous(breaks = c(0,1,5,10,12)) +
  expand_limits(x = c(NA,12)) +
  labs(x = 'rt (min)', caption = 'dashes = trials. red line is 10 minutes', title = 'Retrieval')
```

## Criterion 2: Cued Recall Performance > 3 SDs

Cued recall performance in greater than 3 SDs away from sample mean.

```{r}
df.graded %>%
  dplyr::select(subject_id, session, trial_index, encTrialNum, ends_with('Correct')) %>%
  pivot_longer(cols = ends_with('Correct'), names_to = 'position', values_to = 'correct') %>%
  filter(!is.na(correct)) %>%
  group_by(subject_id, session) %>%
  summarise(prop_correct = mean(correct), .groups = 'drop') %>%
  # z score within session (session)
  group_by(session) %>%
  mutate(zscore = as.double(scale(prop_correct))) %>%
  mutate(exclude = abs(zscore) > 3) %>%
  arrange(desc(abs(zscore))) %>%
  rmarkdown::paged_table()
```

Zscore within session. See code.

## Criterion 3: RT on Surveys < 3 SDs

Average RT to complete the debriefing, SAM, IRI, or VVIQ is faster than 3 SDs away from sample mean.  

```{r}
df.sam %>%
  dplyr::select(subject_id, rt) -> df.sam

df.iri %>%
  dplyr::select(subject_id, rt) -> df.iri

df.vviq %>%
  dplyr::select(subject_id, rt) -> df.vviq

bind_rows(df.sam, df.iri, df.vviq, .id = 'survey') %>%
  mutate(survey = factor(survey, labels = c('sam', 'iri', 'vviq'))) %>%
  # zscore within survey
  group_by(survey) %>%
  mutate(zscore = as.double(scale(rt))) %>%
  mutate(exclude = zscore < -3) %>%
  arrange(desc(abs(zscore))) %>%
  rmarkdown::paged_table()
```
