==> analysis/07_recall_performance.Rmd <==
---
title: "Recall Performance"
date: "`r Sys.Date()`"
author: "Kyle Kurkela"
output: 
  html_document: 
    toc: yes
    toc_float: yes
    code_folding: hide
    df_print: paged
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)

library(tidyverse)
library(ggbeeswarm)
library(ggdist)
library(lmerTest)
library(report)

graded.df <- read_rds('tidy_data/graded_df.rds')
dependency.df <- read_rds('tidy_data/tidy_dependency.rds')
```

## How well did participants do on the cued recall?

```{r}
graded.df %>%
  dplyr::select(subject_id, study_id, session_id, 
                session, condition, encTrialNum, 
                trial_index, ends_with('Correct')) %>%
  pivot_longer(cols = ends_with('Correct'), names_to = 'position', values_to = 'correct') %>%
  filter(!is.na(correct)) %>%
  group_by(subject_id, condition, session) %>%
  summarise(across(correct, mean), .groups = 'drop')-> performance.df

performance.df %>%
  arrange(session, correct) %>%
  rmarkdown::paged_table()
```

```{r}
ggplot(performance.df, aes(y = correct, x = condition)) +
 geom_beeswarm(alpha = 0.5) +
 geom_line(aes(group = subject_id), alpha = 0.5) +
 facet_grid(~session) +
 stat_summary(geom = 'crossbar', width = 0.2, fun.data = mean_se, fill = 'white') +
 labs(y = '% Cued Recall Responses Correct', title = 'Overall Accuracy', x = 'Triad Type') +
 theme_classic() -> plotA

ggsave(filename = 'figures/Figure1.png', 
       plot = plotA, 
       width = 7, height = 5, dpi = 600,
       device = 'png', units = 'in')

plotA
```

```{r}
graded.df %>%
  dplyr::select(subject_id, study_id, session_id, session, condition, encTrialNum, trial_index, ends_with('Correct')) %>%
  pivot_longer(cols = ends_with('Correct'), names_to = 'position', values_to = 'correct') %>%
  filter(!is.na(correct)) -> df
```

```{r}
glmer(correct ~ session * condition + (condition*session|subject_id), 
      family = 'binomial', data = df) -> model1.fit

summary(model1.fit)

report(model1.fit, effectsize_method = 'sdy') -> rep1

summary(rep1)
```

## Dependency

```{r}
dependency.df %>% 
  group_by(subject_id) %>% 
  summarise(across(starts_with('joinedRetrieval'), mean), .groups = 'drop') -> tmp.df

t.test(tmp.df$joinedRetrieval.data, tmp.df$joinedRetrieval.indep, paired = T) -> testResult

report(testResult) %>% 
  summary()
```

```{r}
ggplot(dependency.df, aes(x = condition, y = dependency)) +
  geom_point(alpha = 0.4) +
  geom_line(aes(group = subject_id), alpha = 0.4) +
  stat_summary(geom = 'crossbar', fun.data = mean_se, width = 0.2, fill = 'white') +
  scale_y_continuous(limits = c(NA, 0.5)) +
  theme_classic() +
  labs(title = 'Retrieval Dependency', y = 'Dependency', x = 'Triad Type') +
  guides(color = "none") +
  facet_wrap(~session, ncol = 2) -> plotB

ggsave(filename = 'figures/Figure2.png', 
       plot = plotB, 
       width = 7, height = 5, dpi = 600,
       device = 'png', units = 'in')

plotB
```

```{r}
lmer(dependency ~ session * condition + (1|subject_id), data = dependency.df) -> model1.fit

summary(model1.fit)
```

```{r}
report(model1.fit) -> rep1

summary(rep1)
```

```{r}
dependency.df %>% 
  dplyr::select(subject_id, session, condition, dependency) %>% 
  pivot_wider(values_from = dependency, names_from = condition) %>% 
  mutate(personType = case_when(`famous person` > `famous place` ~ 'person favorer', 
                                `famous person` < `famous place` ~ 'place favorer', 
                                TRUE ~ 'neither')) %>% 
  pivot_longer(cols = all_of(c('famous person', 'famous place')), 
                             names_to = 'condition', 
                             values_to = 'dependency') %>%
  ggplot(., aes(x = condition, y = dependency, color = personType, group = subject_id)) +
  geom_point() +
  geom_line(aes(group = subject_id)) +
  scale_y_continuous(limits = c(NA, 0.5)) +
  theme_light() +
  labs(title = 'Dependency', subtitle = 'Highlighting Person Types', caption = 'color = personType') +
  facet_wrap(~session)
```
