==> analysis/10_across_subjects_correlations.Rmd <==
---
title: "Across Subjects Correlations"
date: "`r Sys.Date()`"
author: "Kyle Kurkela"
output: 
  html_document:
    toc: yes
    toc_float: yes
    code_folding: hide
    df_print: paged    
---

```{r setup, echo=FALSE}
# requirements
suppressPackageStartupMessages(library(tidyverse))
suppressPackageStartupMessages(library(corrplot))

# data
# -- bias
tidy.bias       <- read_rds(file = 'tidy_data/tidy_bias.rds') %>% ungroup()
# -- surveys
tidy.sam        <- read_rds(file = 'tidy_data/tidy_sam.rds')
tidy.iri        <- read_rds(file = 'tidy_data/tidy_iri.rds')
tidy.vivq       <- read_rds(file = 'tidy_data/tidy_vviq.rds')
# -- encoding success
tidy.enc        <- read_rds(file = 'tidy_data/tidy_enc.rds')
# -- dependency
tidy.dependency <- read_rds(file = 'tidy_data/tidy_dependency.rds')
# -- retrieval success
tidy.ret        <- read_rds(file = 'tidy_data/graded_df.rds')

```

```{r tidy_chunk, echo=FALSE}
# minor tidying

tidy.vivq %>%
  rename(visual_imagery = response) -> tidy.vivq

tidy.enc %>%
  group_by(subject_id) %>%
  summarise(success = mean(response_slider)) -> tidy.enc

tidy.dependency %>%
  group_by(subject_id) %>%
  summarise(dependency = mean(dependency)) -> tidy.dependency

tidy.ret %>%
  pivot_longer(cols = ends_with('Correct'), names_to = 'type', values_to = 'isCorrect') %>%
  filter(!is.na(isCorrect)) %>%
  group_by(subject_id) %>%
  summarise(retrieval_success = mean(isCorrect), .groups = 'drop') -> tidy.ret

left_join(tidy.bias, tidy.sam, by = 'subject_id') %>%
  left_join(tidy.iri, by = 'subject_id') %>%
  left_join(tidy.vivq, by = 'subject_id') %>%
  left_join(tidy.enc, by = 'subject_id') %>%
  left_join(tidy.dependency, by = 'subject_id') %>%
  left_join(tidy.ret, by = 'subject_id') -> df
```

```{r echo = FALSE}
# analysis
df %>%
  ungroup() %>%
  dplyr::select(-subject_id) %>%
  corrr::correlate(quiet = T) -> corr_mat

df %>%
  ungroup() %>%
  dplyr::select(-subject_id) %>%
  corrplot::cor.mtest() %>%
  magrittr::extract2('p') %>%
  `diag<-`(1) %>%
  corrr::as_cordf() %>%
  corrr::shave(upper = F) %>%
  corrr::stretch(na.rm = T) %>%
  rename(p = r) -> p

corr_mat %>%
  corrr::shave(upper = F) %>%
  corrr::stretch(na.rm = T) %>%
  left_join(p, by = c('x','y')) -> corr_df
```

## Raw Correlation Matrix

```{r, echo=FALSE,warning=FALSE,message=FALSE}
corr_df %>%
  ggplot(aes(x = x, y = y, fill = r)) +
  geom_raster() +
  geom_text(aes(label = round(r, 2)), color = 'black', size = 8/.pt) +
  scale_fill_gradient2() +
  scale_x_discrete(limits = rev(corr_mat$term)) +
  scale_y_discrete(limits = rev(corr_mat$term)) +
  guides(fill = 'none') +
  theme_classic() +
  theme(axis.text.x = element_text(angle = 90, hjust = 1), axis.title = element_blank())
```

## Correlation Matrix Highlighting p < 0.05

```{r, echo=FALSE,warning=FALSE,message=FALSE}
corr_df %>%
  ggplot(aes(x = x, y = y, fill = r, alpha = p < 0.05)) +
  geom_raster() +
  geom_text(aes(label = round(r, 2)), color = 'black', size = 8/.pt) +
  scale_fill_gradient2() +
  scale_x_discrete(limits = rev(corr_mat$term)) +
  scale_y_discrete(limits = rev(corr_mat$term)) +
  theme_classic() +
  guides(fill = 'none', alpha = 'none') +
  theme(axis.text.x = element_text(angle = 90, hjust = 1), axis.title = element_blank())
```

where:

| variable | description | 
| ----     | ----------  |
`famous person` | average dependency on famous person triads. |
`famous place` | average dependency on famous place triads. |
`bias` | `famous place` - `famous person`. |
`episodic` | sum total score on SAM-episodic subscale. |
`future` | sum total score on SAM-episodic subscale. |
`semantic` | sum total score on SAM-episodic subscale. |
`spatial` | sum total score on SAM-episodic subscale. |
`EC` | sum total score on IRI: Empathetic Concern subscale. "the tendency to experience feelings of sympathy and compassion for unfortunate others" |
`FS` | sum total score on IRI: Fantasy subscale. "the tendency to imaginatively transpose oneself into fictional situations" |
`PD` | sum total score on IRI: Personal Distress subscale. "the tendency to experience distress and discomfort in response to extreme distress in others" |
`PT` | sum total score on IRI: Perspective Taking subscale. "the tendency to spontaneously adopt the psychological point of view of others in everyday life" |
`visual_imagery` | sum total score on the VVIQ. |
| `sucess` | Average Imagintion Success from Encoding. |
| `dependency` | Average Dependency. |
| `retrieval_success` | % of cued recall prompts answered correctly. |

## Retrieval Success and Dependency

```{r}
ggplot(df, aes(x = retrieval_success, y = dependency)) +
  geom_point() +
  geom_smooth(method = 'lm', formula = y ~ x + I(x^2)) +
  labs(title = 'Dependency is a Quadratic Function of Overall Accuracy', 
       caption = 'dots = subjects',
       x = 'Overall Accuracy',
       y = 'Dependency') +
  theme_classic() -> plotA

ggsave(filename = 'figures/Figure3.png', 
       plot = plotA, 
       width = 7, height = 5, dpi = 600,
       device = 'png', units = 'in')

plotA
```

```{r}
lm(dependency ~ retrieval_success + I(retrieval_success^2), data = df) -> model.fit

summary(model.fit)
```
```{r}
report(model.fit) %>%
  summary()
```


## Encoding Success and Dependency

```{r}
ggplot(df, aes(x = dependency, y = success)) +
  geom_point() +
  geom_smooth(method = 'lm', formula = y ~ x) +
  labs(title = 'No Relationship Between Imagination Success and Dependency',
       subtitle = 'Across Subjects',
       caption = 'dots = subjects',
       y = 'Avg Imagination Success at Encoding',
       x = 'Avg Dependency')
```

## Encoding Imagination Success and Retrieval Success

```{r}
ggplot(df, aes(y = retrieval_success, x = success)) +
  geom_point() +
  geom_smooth(method = 'lm', formula = y ~ x) +
  labs(title = 'No Relationship Between Imagination Success and Dependency',
       subtitle = 'Across Subjects',
       caption = 'dots = subjects',
       x = 'Avg Imagination Success at Encoding',
       y = '% Cued Recall Prompts Correct')
```
