==> analysis/10_explore_success.Rmd <==
---
title: "Encoding Success Analysis"
date: "`r Sys.Date()`"
author: "Kyle Kurkela"
output:
  html_document:
    toc: yes
    toc_float: yes
    code_folding: hide
    df_print: paged
---

```{r setup, message=FALSE,warning=FALSE, echo=FALSE}
# requirements
library(tidyverse)
library(ggbeeswarm)
library(ppcor)
library(lmerTest)
library(ggdist)

# data
tidy.enc     <- read_rds('tidy_data/tidy_enc.rds')
```

```{r tidy, echo = FALSE}
# long --> wide
tidy.enc %>%
  dplyr::select(subject_id, session, condition, response_slider) %>%
  group_by(subject_id, session, condition) %>%
  summarise(across(.cols = response_slider, .fns = mean), .groups = 'drop') -> df

df %>%
  pivot_wider(id_cols = all_of(c('subject_id', 'session')), names_from = condition, values_from = response_slider) %>%
  group_by(subject_id) %>%
  mutate(person_prefer = `famous person` > `famous place`,
         place_prefer = `famous person` < `famous place`) %>%
  mutate(type = case_when(sum(person_prefer) == 2 ~ 'strong person bias',
                          sum(place_prefer) == 2 ~ 'strong place bias',
                          TRUE ~ 'ambiguous')) -> figureData

tidy.enc %>%
  dplyr::select(subject_id, session, condition, response_slider) %>%
  group_by(subject_id, condition) %>%
  summarise(across(response_slider, .fns = mean), .groups = 'drop') %>%
  pivot_wider(id_cols = all_of(c('subject_id')), names_from = condition, values_from = response_slider) -> summarised.df
```

## Encoding Success is Biased Towards Persons in Place/Person space

In the animation below, I slowly build up a figure placing our subjects into "Place/Person" space. X-axis = dependency on famous person triads, Y-axis = dependency on famous place triads (Figure 1). You can divide this space on the diagonal (Figure 2) into "place-bias" (upper left; red; Figure 3) and "person-bias" (lower right; blue; Figure 4) zones. The asterisks represent subjects average across sessions (Figure 5). The "dumbells" (i.e., line with two dots at either end) also represents each subject -- the end of each "dumbell" is how the participant performed in session 1 and session 2. The asterisks -- representing each subject's average across sessions -- sits in the center of this line (Figure 6). Participant's "dumbells" are then color coded based on whether they had a `strong place bias` (red), a `strong person bias` (blue), or it was `ambiguous` (black) whether or not they have a bias (Figure 7).


```{r, animation.hook='ffmpeg', echo=FALSE, interval=2, fig.align='center'}
(ggplot(figureData, aes(x = `famous person`, y = `famous place`)) +
    guides(color = 'none') +
    theme_minimal() +
    scale_x_continuous(limits = c(0, 100)) +
    scale_y_continuous(limits = c(0, 100)) +
    theme(aspect.ratio = 1) -> blank)

(blank + geom_abline(slope = 1, intercept = 0, linetype = 'solid') -> blank)

(blank + 
    annotate(geom = 'polygon', x = c(0,0,100), y = c(0,100,100), alpha = 0.3, fill = scales::muted('red')) +
    annotate(geom = 'label', x = 25, y = 90, label = '"place" bias', color = scales::muted('red')) -> blank)

(blank +
    annotate(geom = 'polygon', x = c(0,100,100), y = c(0,0,100), alpha = 0.3, fill = scales::muted('blue')) +
    annotate(geom = 'label', x = 90, y = 25, label = '"person" bias', color = scales::muted('blue')) -> blank)

blank + 
  geom_point(data = summarised.df, shape = 'asterisk', size = 2)

blank +
  geom_point(data = summarised.df, shape = 'asterisk', size = 2) +
  geom_point(aes()) + 
  geom_path(aes(group = subject_id), linetype = 'solid')

blank + 
  geom_point(data = summarised.df, shape = 'asterisk', size = 2) +
  geom_point(aes(color = type)) + 
  geom_path(aes(group = subject_id, color = type)) + 
  scale_color_manual(values = c('black', 'blue', 'red'))
blank + 
  geom_point(data = summarised.df, shape = 'asterisk', size = 2)
```

A table counting the number of subjects with a `strong place bias` (i.e., place bias in BOTH sessions), `strong person bias` (i.e., person bias in BOTH sessions), and an ambiguous bias (they "jumped" from session to session OR they perfectly hug the diagonal):  

```{r}
figureData %>%
  ungroup() %>%
  count(subject_id, type) %>%
  count(type)
```

## Person success > Place success

```{r echo = FALSE}
tidy.enc %>%
  dplyr::select(subject_id, session, condition, response_slider) %>%
  group_by(subject_id, session, condition) %>%
  summarise(across(.cols = response_slider, .fns = mean), .groups = 'drop') -> df
```

### Visual

```{r}
ggplot(df, aes(x = condition, y = response_slider)) +
  geom_point() +
  geom_line(aes(group = subject_id), alpha = 0.2) +
  stat_summary(geom = 'crossbar', fun.data = mean_se, width = 0.2, fill = 'gray') +
  facet_grid(~session) +
  theme_classic()
```

### Stats

```{r}
lmer(response_slider ~ session * condition + (session|subject_id), data = tidy.enc) -> model1.fit

summary(model1.fit)
```

## What about our key hypothetical test?

```{r echo = FALSE}
tidy.enc %>%
  dplyr::select(subject_id, session, condition, response_slider) %>%
  group_by(subject_id, session, condition) %>%
  summarise(across(.cols = response_slider, .fns = mean), .groups = 'drop') %>%
  unite(col = 'sess_cond', session, condition) %>%
  pivot_wider(id_cols = all_of(c('subject_id')), names_from = sess_cond, values_from = response_slider) -> df
```

`session1_famous person` ~~ `session2_famous person` CONTROLLING FOR `session1_famous place`, `session2_famous place`: 

```{r, echo=TRUE}
(ppcor::pcor.test(x = df$`session1_famous person`, 
                 y = df$`session2_famous person`, 
                 z = df[,c('session1_famous place', 'session2_famous place')]) -> result1)
```

`session1_famous place` ~~ `session2_famous place` CONTROLLING FOR `session1_famous person`, `session2_famous person`: 

```{r}
(ppcor::pcor.test(x = df$`session1_famous place`, 
                 y = df$`session2_famous place`, 
                 z = df[,c('session1_famous person', 'session2_famous person')]) -> result2)
```

What is our stopping decision?

```{r}
# stopping decision

theMaxPvalue <- max(result1$p.value, result2$p.value)

alphaStrong <- 0.035
alphaWeak   <- 0.282

if(theMaxPvalue < alphaStrong){
  print('Reject the Null Hypothesis')
}

if(theMaxPvalue > alphaWeak){
  print('Fail to Reject the Null Hypothesis')
}

if(theMaxPvalue > alphaStrong & theMaxPvalue < alphaWeak){
  print('Inconclusive. Continue data collection.')
}
```

## Sanity check -- encoding success predicts later memory?

```{r, echo=FALSE, warning=FALSE, message=FALSE}
graded.df <- read_rds('tidy_data/graded_df.rds')

left_join(graded.df, tidy.enc) %>%
  dplyr::select(subject_id, condition, session, trial_index, encTrialNum, ends_with('Correct'), response_slider) %>%
  pivot_longer(cols = ends_with('Correct'), names_to = 'type', values_to = 'isCorrect') %>%
  filter(!is.na(isCorrect)) -> df
```

### Stats

```{r, warning=FALSE,message=FALSE}
glmer(isCorrect ~ 1 + (1|subject_id), data = df) -> model0.fit
glmer(isCorrect ~ response_slider + (1|subject_id), data = df) -> model1.fit

anova(model0.fit, model1.fit)
```

```{r}
summary(model1.fit)
```


### Visual

```{r}
predict(model1.fit, type = 'response') -> df$predictedProbCorrect
predict(model1.fit, type = 'response', re.form = NA) -> df$predictedProbCorrectFixed

ggplot(df, aes(x = response_slider, y = as.numeric(isCorrect))) +
  stat_histinterval(orientation = 'y', aes(side = if_else(!isCorrect, 'bottom', 'top'), justification = if_else(!isCorrect, 1.1, -0.1) )) +
  geom_line(aes(group = subject_id, y = predictedProbCorrect, color = subject_id), alpha = 0.2) +  
  geom_line(aes(y = predictedProbCorrectFixed), color = 'black', linewidth = 2) +
  guides(color = 'none') +
  labs(title = 'Success Ratings At Encoding Predict Retrieval Success',
       y = 'Retrieval Success (Probability)',
       x = 'Self-Reported Imaination Success at Encoding',
       caption = 'histograms display counts at different bins of success')
```
