==> analysis/10_familiarity_ratings.Rmd <==
---
title: "Familarity Ratings Analysis"
date: "`r Sys.Date()`"
author: "Kyle Kurkela"
output:
  html_document:
    toc: yes
    toc_float: yes
    code_folding: hide
    df_print: paged
---

```{r setup, message=FALSE,warning=FALSE, echo=FALSE}
# requirements
library(tidyverse)
library(ggbeeswarm)
library(ppcor)
library(lmerTest)
library(ggdist)

# data
df <- read_rds('tidy_data/compiled_experiment.rds')

extract_fam_ratings <- function(x){
  # x is a json formatted string.

  jsonlite::parse_json(x) %>% 
    unlist() %>% 
    as_tibble_row() %>% 
    pivot_longer(cols = everything(), names_to = 'stim', values_to = 'rating') %>% 
    separate(col = 'stim', into = c('stimType', 'stimuli'), sep = '(?<=[es])(\\.)')
}
```

```{r}
df %>%
  filter(phase == 'stim_ratings') %>% 
  dplyr::select(where(~!all(is.na(.x)))) %>%
  dplyr::select(subject_id, session, version, response) -> df

# only subjects with both sessions
df %>%
  nest(data = -all_of(c('subject_id', 'session'))) %>%
  pivot_wider(id_cols = subject_id, names_from = session, values_from = data) %>%
  filter(map_lgl(session1, is_tibble) & map_lgl(session2, is_tibble)) %>%
  pivot_longer(cols = starts_with('session'), names_to = 'session', values_to = 'data') %>%
  unnest(cols = data) -> df

df %>%
  mutate(response = map(.x = response, .f = extract_fam_ratings)) %>%
  unnest(response) -> fam.ratings.df
```

## Does familarity rating differ between conditions?

Yes. Persons >> Places. See below:

```{r, echo = FALSE}
fam.ratings.df %>%
  group_by(subject_id, session, stimType) %>%
  summarise(rating = mean(rating), .groups = 'drop') -> df
```

### Visuals

Person/Place Space

```{r}
df %>% 
  pivot_wider(id_cols = all_of(c('subject_id', 'session')), 
              names_from = stimType, values_from = rating) %>%
  ggplot(aes(x = people, y = places)) +
  geom_abline(slope = 1, intercept = 0) +
  geom_point() +
  scale_x_continuous(limits = c(0, 4)) +
  scale_y_continuous(limits = c(0, 4)) +
  geom_line(aes(group = subject_id)) +
  labs(x = 'Avg Famous Person Familarity', 
       y = 'Avg Famous Places Familarity', 
       caption = 'dumbells = subjects. Dots = sessions', 
       title = 'Person Familarity >> Place Familarity')
```

```{r}
ggplot(df, aes(x = stimType, y = rating)) +
  geom_point(alpha = 0.4) +
  geom_line(aes(group = subject_id), alpha = 0.4) +
  stat_summary(geom = 'crossbar', fun.data = mean_se, width = 0.2, fill = 'white') +
  facet_grid(~session) +
  labs(title = 'Familarity of Famous Experimental Stimuli', 
       x = 'Famous Stimuli Type', y = 'Familarity Rating', 
       caption = 'dots = subjects. Lines connect subjects.') +
  theme_classic() -> plt

ggsave(filename = 'figures/Figure17.png', 
       plot = plt, device = 'png', 
       width = 7, height = 5, units = 'in', dpi = 600)

plt
```

```{r}
lmer(data = df, rating ~ session*stimType + (stimType|subject_id)) -> model.fit
summary(model.fit)
```

```{r}
report::report(model.fit)
```


## Does familarity rating predict encoding success?

YES.

```{r, echo = FALSE}
df <- read_rds('tidy_data/tidy_enc.rds')

left_join(df, fam.ratings.df, by = c('subject_id', 'session', 'key' = 'stimuli')) -> df
```

### Stats

```{r}
lmer(response_slider ~ rating + (rating|subject_id), data = df) -> model.1
summary(model.1)
```

### Visual

```{r}
ggplot(df, aes(x = rating, y = response_slider)) +
  geom_quasirandom(alpha = 0.2, width = 0.2) +
  stat_summary(fun.data = mean_se, geom = 'crossbar', width = 0.2, fill = 'red', color = 'blue') +
  guides(color = 'none') +
  labs(y = 'Encoding Imagination Success', 
       x = 'Familarity Rating for Key Stimulus', 
       caption = 'dots = trials', 
       title = 'Familarity Rating Predicts Imagination Success')
```

## Does familarity rating predict retrieval success?

YES.

```{r, echo = FALSE}
df <- read_rds('tidy_data/graded_df.rds')

left_join(df, fam.ratings.df, by = c('subject_id', 'session', 'key' = 'stimuli')) %>%
  pivot_longer(cols = ends_with('Correct'), names_to = 'type', values_to = 'isCorrect') %>%
  filter(!is.na(isCorrect)) -> df

# assign each encoding trial a unique id to use as a nested random effect
df %>%
  nest(cols = -c(subject_id, session, encTrialNum)) %>%
  arrange(subject_id, session, encTrialNum) %>%
  mutate(encTrialID = rep(1:48, 60)) %>%
  unnest(cols) -> df
```

### Stats

```{r}
library(lavaan)

model <- '
  rating ~ condition
  isCorrect ~ condition
  isCorrect ~ rating
'

sem(model = model, data = df, ordered = c('isCorrect', 'condition')) -> model.fit

summary(model.fit)
```
```{r}
df %>%
  dplyr::select(subject_id, session, condition, rating, isCorrect) %>%
  mutate(subject_id = factor(subject_id, labels = 1:60)) %>%
  mutate(session = factor(session, levels = c('session1', 'session2'), labels = c(1,2))) %>%
  mutate(condition = factor(condition, levels = c('famous place', 'famous person'), labels = c(1,2))) %>%
  mutate(isCorrect = as.double(isCorrect)) %>%
  write_delim(., file = 'test.txt', col_names = FALSE)
```


```{r}
glmer(isCorrect ~ condition * session + rating +
        (condition*rating|subject_id), 
      data = df, family = 'binomial') -> model.1

glmer(isCorrect ~ rating + 
        (condition*rating|subject_id),
      data = df, family = 'binomial') -> model.3

anova(model.1, model.3)
```

```{r}
report::report(model.1) -> res1

res1
```


```{r}
df$predictionRE <- predict(model.1, newdata = df, type = 'response', re.form = NULL)
df$predictionFE <- predict(model.1, newdata = df, type = 'response', re.form = NA)

df %>%
  group_by(subject_id, encTrialID) %>%
  summarise(isCorrect = mean(isCorrect),
            session = unique(session),
            condition = unique(condition),
            rating = unique(rating),
            predictionRE = unique(predictionRE),
            predictionFE = unique(predictionFE), .groups = 'drop') -> summ.df
```

```{r}
ggplot(summ.df, aes(x = rating, y = isCorrect)) +
  geom_point() +
  geom_line(aes(y = predictionFE)) +
  guides(color = 'none') +
  facet_grid(session~condition)
```

```{r}
ggplot(summ.df, aes(x = rating, y = isCorrect)) +
  geom_point() +
  geom_line(aes(y = predictionFE)) +
  guides(color = 'none') +
  facet_grid(session~condition)
```


### Visual

```{r}
ggplot(df, aes(x = rating, y = numCorrect)) +
  geom_quasirandom(alpha = 0.2, width = 0.2) +
  stat_summary(fun.data = mean_se, geom = 'crossbar', width = 0.2, fill = 'red', color = 'blue') +
  guides(color = 'none') +
  labs(y = 'Number of Cued Retrievals Answered Correctly', 
       x = 'Familarity Rating for Key Stimulus', 
       caption = 'dots = trials', 
       title = 'Familarity Rating Predicts Retrieval Accuracy')
```

## Does familarity rating predict dependency?

NO.

```{r, echo = FALSE}
df <- read_rds('tidy_data/tidy_dependency.rds')

fam.ratings.df %>%
  group_by(subject_id, session, stimType) %>%
  summarise(rating = mean(rating), .groups = 'drop') %>%
  mutate(stimType = case_when(stimType == 'people' ~ 'famous person',
                              stimType == 'places' ~ 'famous place')) -> fam.ratings.summarised

left_join(df,fam.ratings.summarised, by = c('subject_id', 'session', 'condition' = 'stimType')) -> df
```

### Stats

```{r}
lmer(dependency ~ rating + (rating|subject_id), data = df) -> model.1
summary(model.1)
```

### Visual

```{r}
df$pred <- predict(model.1)
df$pred_fix <- predict(model.1, re.form = NA)

ggplot(df, aes(x = rating, y = dependency)) +
  geom_point() +
  geom_line(aes(y = pred, color = subject_id), alpha = 0.4) +
  geom_line(aes(y = pred_fix), color = 'blue', linewidth = 2) +
  guides(color = 'none') +
  labs(y = 'Retrieval Dependency', 
       x = 'Average Familarity Does Not Predict Retrieval Dependency', 
       caption = 'dots = subject:session:condition.\nFadded color lines = random effects model fit for subjects.\nThick blue line = fixed effect estimate', 
       title = 'Familarity Rating Predicts Imagination Success')
```

## Does familarity rating predict bias?

```{r}
df.bias <- read_rds('tidy_data/tidy_bias.rds')

head(df.bias)
```

```{r}
head(fam.ratings.df)
```

```{r}
fam.ratings.df %>%
  group_by(subject_id, stimType) %>%
  summarise(rating = mean(rating), .groups = 'drop') %>%
  pivot_wider(id_cols = subject_id, names_from = stimType, values_from = rating) %>%
  mutate(bias_fam = places - people) -> fam.ratings.bias.df
```

### Are the familarity Ratings Biased?

"Bias" is the distance away from the diagonal line:

```{r}
ggplot(fam.ratings.bias.df, aes(x = people, y = places)) +
  geom_point() +
  geom_abline(slope = 1, intercept = 0) +
  scale_x_continuous(limits = c(0,4)) +
  scale_y_continuous(limits = c(0,4)) +
  labs(title = 'People Familarity >> Place Familarity', x = 'Avg Famous Person Familarity', y = 'Avg Famous Place Familarity')
```

### Does bias predict bias?

```{r}
left_join(df.bias, fam.ratings.bias.df, by = 'subject_id') -> df.bias

lm(data = df.bias, bias ~ bias_fam) -> model.fit

summary(model.fit)
```

```{r}
ggplot(df.bias, aes(x = bias_fam, y = bias)) +
  geom_point() +
  geom_smooth(method = 'lm', formula = y ~ x) +
  labs(x = 'avg famous place familarity - avg famous person familarity',
       y = 'Bias = place dependency - person dependency',
       caption = 'dots = subjects', 
       title = 'Bias is NOT predicted by relative familarity of famous places/famous people.')
```
