==> analysis/11_tidy_bds.Rmd <==
---
title: "explore_bds"
author: "Kyle Kurkela"
date: "2023-08-16"
output: html_document
---

```{r setup, include=FALSE}
library(tidyverse)
source('functions/extract_corr_answer.R')

df.bds <- read_rds('tidy_data/tidy_bds.rds')
```

## tidy

```{r}
df.bds %>%
  mutate(corr_answer = map_chr(bds_presentation, extract_corr_answer)) %>%
  mutate(isCorrect = map2_lgl(response, corr_answer, str_equal)) %>%
  group_by(subject_id) %>%
  summarise(across(isCorrect, c('prop' = mean, 'count' = sum, tot = length))) -> graded.bds

write_rds(x = graded.bds, file = 'tidy_data/graded_bds.rds')
```
